+++ src/parsers/HTTPLexer.cpp
@@ -1330,7 +1330,7 @@ try
             more();
         }
     }
-    else if (matched().length() - 2 == d_chunked_length) {
+    else if (matched().length() - 2 == (size_t) d_chunked_length) {
         d_chunked.insert(d_chunked.length(), matched());
         muzzley::trim(d_chunked);
         setMatched("");
